+++ run/api/index.js
@@ -18,6 +18,7 @@ const erc721Tokens = require('./erc721Tokens');
 const status = require('./status');
 const external = require('./external');
 const domains = require('./domains');
+const faucets = require('./faucets');
 
 router.use('/blocks', blocks);
 router.use('/contracts', contracts);
@@ -36,6 +37,7 @@ router.use('/erc721Tokens', erc721Tokens);
 router.use('/status', status);
 router.use('/external', external);
 router.use('/domains', domains);
+router.use('/faucets', faucets);
 
 if (isDemoEnabled()) {
     const demo = require('./demo');