+++ src/App.vue
@@ -102,6 +102,15 @@
                     </v-list-item-content>
                 </v-list-item>
 
+                <v-list-item link :to="'/faucet'" v-if="isPublicExplorer && publicExplorer.faucet">
+                    <v-list-item-icon>
+                        <v-icon>mdi-faucet</v-icon>
+                    </v-list-item-icon>
+                    <v-list-item-content>
+                        <v-list-item-title>Faucet</v-list-item-title>
+                    </v-list-item-content>
+                </v-list-item>
+
                 <v-list-item link :to="'/status'" v-if="(isUserAdmin && currentWorkspace.public) || currentWorkspace.statusPageEnabled">
                     <v-list-item-icon>
                         <v-icon>mdi-heart-circle</v-icon>