+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.36.1] - 2024-06-12
+### Fixed
+- Bug preventing explorer from loading when using a custom domain
+- Bug with custom domains status request
+
+## [3.36.0] - 2024-06-12
+### Added
+- Faucet: explorers can integrate a faucet for their users. Enter the request interval & the drip amount, fill your assigned address with your tokens, and your faucet is ready! It comes with analytics, transaction history, address history.
+
+## [3.35.6] - 2024-06-05
+### Fixed
+- Bug preventing transaction rollback if an event is not inserted properly
+
+## [3.35.5] - 2024-05-21
+### Fixed
+- Receipt re-enqueing bug
+
 ## [3.35.4] - 2024-05-28
 ### Changed
 - Cleaner skeleton loading